feat: add collision handling for ships & lasers
The player's lasers should damage the enemy's armor on collision and the
enemy should damage the player's armor on collision.

I also removed the distance check for the enemy ship as it is trying to
run into the player to damage them.

Also added more enemies to the room so the player can actually die, and
made sure that the enemy ship inherited from the ship parent object.